+++ hpx/runtime/components/server/locking_hook.hpp
@@ -8,9 +8,9 @@
 
 #include <hpx/hpx_fwd.hpp>
 #include <hpx/runtime/get_lva.hpp>
+#include <hpx/runtime/threads/coroutines/coroutine.hpp>
 #include <hpx/util/unlock_guard.hpp>
 #include <hpx/util/move.hpp>
-#include <hpx/runtime/coroutine/coroutine.hpp>
 #include <hpx/util/register_locks.hpp>
 #include <hpx/lcos/local/spinlock.hpp>
