+++ hpx/runtime/threads/coroutines/detail/context_generic_context.hpp
@@ -6,17 +6,17 @@
 //  (See accompanying file LICENSE_1_0.txt or copy at
 //  http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef HPX_RUNTIME_COROUTINE_DETAIL_CONTEXT_GENERIC_HPP
-#define HPX_RUNTIME_COROUTINE_DETAIL_CONTEXT_GENERIC_HPP
+#ifndef HPX_RUNTIME_THREADS_COROUTINES_DETAIL_CONTEXT_GENERIC_HPP
+#define HPX_RUNTIME_THREADS_COROUTINES_DETAIL_CONTEXT_GENERIC_HPP
 
 #include <hpx/config.hpp>
 #include <hpx/util/assert.hpp>
-#include <hpx/runtime/coroutine/detail/get_stack_pointer.hpp>
-#include <hpx/runtime/coroutine/exception.hpp>
-#include <hpx/runtime/coroutine/detail/swap_context.hpp>
+#include <hpx/runtime/threads/coroutines/detail/get_stack_pointer.hpp>
+#include <hpx/runtime/threads/coroutines/exception.hpp>
+#include <hpx/runtime/threads/coroutines/detail/swap_context.hpp>
 
 #if defined(_POSIX_VERSION)
-#include <hpx/runtime/coroutine/detail/posix_utility.hpp>
+#include <hpx/runtime/threads/coroutines/detail/posix_utility.hpp>
 #endif
 
 #include <hpx/util/get_and_reset_value.hpp>
@@ -63,7 +63,7 @@ extern "C"
 #endif
 
 ///////////////////////////////////////////////////////////////////////////////
-namespace hpx { namespace coroutines
+namespace hpx { namespace threads { namespace coroutines
 {
     // some platforms need special preparation of the main thread
     struct prepare_main_thread
@@ -289,6 +289,6 @@ namespace hpx { namespace coroutines
 
         typedef fcontext_context_impl context_impl;
     }}
-}}
+}}}
 
-#endif /*HPX_RUNTIME_COROUTINE_DETAIL_CONTEXT_GENERIC_HPP*/
+#endif /*HPX_RUNTIME_THREADS_COROUTINES_DETAIL_CONTEXT_GENERIC_HPP*/
