+++ hpx/runtime/threads/coroutines/detail/get_stack_pointer.hpp
@@ -4,8 +4,8 @@
 //  (See accompanying file LICENSE_1_0.txt or copy at
 //  http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef HPX_RUNTIME_COROUTINE_DETAIL_GET_STACK_POINTER_HPP
-#define HPX_RUNTIME_COROUTINE_DETAIL_GET_STACK_POINTER_HPP
+#ifndef HPX_RUNTIME_THREADS_COROUTINES_DETAIL_GET_STACK_POINTER_HPP
+#define HPX_RUNTIME_THREADS_COROUTINES_DETAIL_GET_STACK_POINTER_HPP
 
 #if !defined(HPX_WINDOWS)
 #if defined(__x86_64__) || defined(__amd64)                                    \
@@ -18,7 +18,7 @@
 #include <cstddef>
 #include <limits>
 
-namespace hpx { namespace coroutines { namespace detail
+namespace hpx { namespace threads { namespace coroutines { namespace detail
 {
     inline std::size_t get_stack_ptr()
     {
@@ -35,8 +35,8 @@ namespace hpx { namespace coroutines { namespace detail
 #endif
         return stack_ptr;
     }
-}}}
+}}}}
 
 #endif
 
-#endif /*HPX_RUNTIME_COROUTINE_DETAIL_GET_STACK_POINTER_HPP*/
+#endif /*HPX_RUNTIME_THREADS_COROUTINES_DETAIL_GET_STACK_POINTER_HPP*/
