+++ hpx/runtime/threads/coroutines/detail/swap_context.hpp
@@ -26,31 +26,19 @@
 //  (See accompanying file LICENSE_1_0.txt or copy at
 //  http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef HPX_RUNTIME_COROUTINE_DETAIL_COROUTINE_ACCESSOR_HPP
-#define HPX_RUNTIME_COROUTINE_DETAIL_COROUTINE_ACCESSOR_HPP
+#ifndef HPX_RUNTIME_THREADS_COROUTINES_DETAIL_SWAP_CONTEXT_HPP
+#define HPX_RUNTIME_THREADS_COROUTINES_DETAIL_SWAP_CONTEXT_HPP
 
-namespace hpx { namespace coroutines { namespace detail
+namespace hpx { namespace threads { namespace coroutines { namespace detail
 {
-    struct coroutine_accessor
-    {
-        template <typename Coroutine>
-        static void acquire(Coroutine& x)
-        {
-            x.acquire();
-        }
+    class default_hint {};
+    class yield_hint : public default_hint {};
+    class yield_to_hint : public default_hint {};
+    class invoke_hint : public default_hint {};
 
-        template <typename Coroutine>
-        static void release(Coroutine& x)
-        {
-            x.release();
-        }
+    /////////////////////////////////////////////////////////////////////////////
+    // This is the base class of all context implementations
+    struct context_impl_base {};
+}}}}
 
-        template <typename Coroutine>
-        static typename Coroutine::impl_ptr get_impl(Coroutine& x)
-        {
-            return x.get_impl();
-        }
-    };
-}}}
-
-#endif /*HPX_RUNTIME_COROUTINE_DETAIL_COROUTINE_ACCESSOR_HPP*/
+#endif /*HPX_RUNTIME_THREADS_COROUTINES_DETAIL_SWAP_CONTEXT_HPP*/
