+++ hpx/runtime/threads/scheduler_specific_ptr.hpp
@@ -14,7 +14,7 @@
 #include <hpx/config.hpp>
 #include <hpx/config/emulate_deleted.hpp>
 #include <hpx/runtime/threads/thread_data_fwd.hpp>
-#include <hpx/runtime/coroutine/detail/tss.hpp>
+#include <hpx/runtime/threads/coroutines/detail/tss.hpp>
 
 #include <boost/shared_ptr.hpp>
 #include <boost/make_shared.hpp>
