+++ hpx/runtime/threads/thread_data_fwd.hpp
@@ -11,22 +11,8 @@
 #include <hpx/config.hpp>
 #include <hpx/exception_fwd.hpp>
 #include <hpx/runtime/threads/thread_enums.hpp>
+#include <hpx/runtime/threads/coroutines/detail/coroutine_impl.hpp>
 #include <hpx/util/unique_function.hpp>
-#include <hpx/runtime/coroutine/detail/coroutine_impl.hpp>
-
-namespace hpx { namespace coroutines
-{
-    /// \cond NOINTERNAL
-    namespace detail
-    {
-        class coroutine_self;
-
-        class coroutine_impl;
-    }
-
-    class coroutine;
-    /// \endcond
-}}
 
 namespace hpx { namespace threads
 {
@@ -44,6 +30,20 @@ namespace hpx { namespace threads
     class HPX_EXPORT executor;
 
     ///////////////////////////////////////////////////////////////////////
+    namespace coroutines
+    {
+        /// \cond NOINTERNAL
+        namespace detail
+        {
+            class coroutine_self;
+
+            class coroutine_impl;
+        }
+
+        class coroutine;
+        /// \endcond
+    }
+
     typedef coroutines::coroutine coroutine_type;
 
     typedef coroutines::detail::coroutine_self thread_self;
