+++ src/runtime/threads/coroutines/detail/tss.cpp
@@ -10,16 +10,16 @@
 
 #include <hpx/config.hpp>
 #include <hpx/hpx_fwd.hpp>
-#include <hpx/runtime/coroutine/coroutine.hpp>
-#include <hpx/runtime/coroutine/detail/coroutine_self.hpp>
-#include <hpx/runtime/coroutine/detail/tss.hpp>
+#include <hpx/runtime/threads/coroutines/coroutine.hpp>
+#include <hpx/runtime/threads/coroutines/detail/coroutine_self.hpp>
+#include <hpx/runtime/threads/coroutines/detail/tss.hpp>
 #include <hpx/util/assert.hpp>
 
 #include <hpx/runtime/threads_fwd.hpp>
 
 #include <map>
 
-namespace hpx { namespace coroutines { namespace detail
+namespace hpx { namespace threads { namespace coroutines { namespace detail
 {
     ///////////////////////////////////////////////////////////////////////////
     void tss_data_node::cleanup(bool cleanup_existing)
@@ -207,5 +207,5 @@ namespace hpx { namespace coroutines { namespace detail
         }
 #endif
     }
-}}}
+}}}}
