+++ src/runtime_impl.cpp
@@ -624,7 +624,7 @@ namespace hpx {
             thread_support_->register_thread(name);
 
             // initialize coroutines context switcher
-            hpx::coroutines::thread_startup(name);
+            hpx::threads::coroutines::thread_startup(name);
 
             // register this thread with any possibly active Intel tool
             HPX_ITT_THREAD_SET_NAME(name);
@@ -661,7 +661,7 @@ namespace hpx {
     void runtime_impl<SchedulingPolicy>::deinit_tss()
     {
         // initialize coroutines context switcher
-        hpx::coroutines::thread_shutdown();
+        hpx::threads::coroutines::thread_shutdown();
 
         // reset applier TSS
         applier_.deinit_tss();
